Let's remove the toast pop-ups when changes are made.
Initiative Tracker: Remove the Description to make more room for the initiative list. When adding enemies, Add an option to add multiple of the same enemy type at once, appending a number to each one in initiative. Additionally, for enemies, give an option to roll their initiative from the add enemy dialogue, with an input for their initiative modifier to be added to their d20 roll

Dice Roller: add an option to clear the roll log, and when advantage is selected,  add a green border to the Roll Button. When Disadvantage is selected, add a red border. The Input field can have the suggestion (e.g., d20+3, 3d6+4) in the field, and if no input, the Roll button can say Roll d20