Adventure Recap Page - Retitle on the page to simply "Adventure Recap".
delete the Campaign Log feature description

If the campaign log gets too long, add a scroll-bar to browse further.

Have the summarizer re-title the session as "Session [x] - [1-4 word summary/'Episode Title'"

Add a button to clear the Campaign Log. Have a confirmation pop-up, then a SECOND pop-up and have the user type the word DELETE in order to confirm the log deletion