Change the header of the Combat Tracker tool to reflect our renaming.
Use a skull icon for the monster mash icon.

Use the same Vertical Exit bar for the Combat Tracker and the Dice Roller.
For the dice roller, When a roll is made, have an empty space populate at the top of the roll log, and then add an animation of a die rolling from the right to the left in that space before showing the result