The DM Tools Drawer doesn't need a DM Tools Header.
The Dice roller should have an animation of a die literally being rolled from the right to the left.

When the user clicks and drags on the Dice Roller Icon, I'd like a dice icon to follow the cursor and spin. Then when the drag is released, I'd like the icon to fall randomly within 100px of the curser and display a Hexagon with a d20 result it in. The result should stay there for 5 seconds (or until the user clicks elsewhere), and then the result should animate as if it were moving into the Dice Roller Tool. This result should appear in the Roll Log and note that it was Quick-Rolled